eatgo-customer-api/src/test/java/kr/co/fastcampus/eatgo/application/ReservationServiceTest.java
package kr.co.fastcampus.eatgo.application;

import kr.co.fastcampus.eatgo.domain.Reservation;
import kr.co.fastcampus.eatgo.domain.ReservationRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.verify;


class ReservationServiceTest {
    private ReservationService reservationService;

    @Mock
    private ReservationRepository reservationRepository;

    @BeforeEach
    public void setUp(){
        MockitoAnnotations.initMocks(this);
        reservationService = new ReservationService(reservationRepository);
    }

    @Test
    public void addReservation(){
        Integer restaurantId = 3;
        Integer userId = 9;
        String name = "정현수";
        String date = "2019-12-24";
        String time = "20:00";
        Integer partySize = 20;

        given(reservationRepository.save(any())).will(invocation -> {
           return invocation.getArgument(0);
        });

        Reservation reservation = reservationService.addReservation(restaurantId, userId, name, date, time, partySize);

        verify(reservationRepository).save(any());
        assertThat(reservation.getName()).isEqualTo("정현수");
    }

}